requirements.txt
Django>=5.0.7
django-widget-tweaks
pandas
Pillow
PyMuPDF
python-pptx 
gunicorn
whitenoise
dj-database-url
psycopg2-binary